move_hand/requirements.txt
numpy==1.24.3
matplotlib
sympy
tqdm
bezier
numpy-quaternion